+++ app/src/main/java/com/shootr/mobile/ui/views/PollResultsView.java
@@ -25,4 +25,6 @@ public interface PollResultsView extends LoadDataView {
   void showSharedPoll();
 
   void hideFooter();
+
+  void showDailyPollText();
 }