+++ app/src/main/java/com/shootr/mobile/ui/views/nullview/NullStreamTimelineView.java
@@ -174,6 +174,10 @@ public class NullStreamTimelineView implements StreamTimelineView {
         /* no-op */
     }
 
+    @Override public void showFollowToast() {
+        /* no-op */
+    }
+
     @Override public void showEmpty() {
         /* no-op */
     }